Removed address schema and merged with user model
