+++ server/app/routes/index.js
@@ -9,7 +9,6 @@ router.use('/products', require('./products'));
 router.use('/reviews', require('./reviews'));
 router.use('/users', require('./users'));
 router.use('/orderitem', require('./orderItems'));
-router.use('/address', require('./address'));
 
 // Make sure this is after all of
 // the registered routes!
